bot/requirements.txt
aiogram>=3.0
python-docx
PyPDF2
python-dotenv